backend/requirements.txt
# Flask and web framework dependencies
Flask==2.3.3
Flask-CORS==4.0.0

# Firebase dependencies
firebase-admin==6.2.0

# Environment and configuration
python-dotenv==1.0.0

# HTTP requests (for AI service integration)
requests==2.31.0

# Data validation and serialization
marshmallow==3.20.1

# Date and time utilities
python-dateutil==2.8.2

# Logging and monitoring
structlog==23.1.0

# Testing dependencies (for development)
pytest==7.4.2
pytest-flask==1.2.0
pytest-cov==4.1.0

# Development dependencies
black==23.7.0
flake8==6.0.0

# Production server (optional)
gunicorn==21.2.0

# Security
cryptography==41.0.4
PyJWT==2.8.0
